requirements.txt
openai>=1.75.0
pyyaml>=6.0
